requirements.txt
pytest
fastapi[all]
requests
opencv-python
matplotlib
pillow
scikit-image
selenium
Scrapy
beautifulsoup4